refactor: util/s3 split into separate files
