+++ .env.template
@@ -4,6 +4,7 @@ NODE_ENV=development
 OPEN_BALENA_S3_URL=https://s3.my-open-balena-server.com
 OPEN_BALENA_S3_ACCESS_KEY=
 OPEN_BALENA_S3_SECRET_KEY=
+OPEN_BALENA_S3_REGISTRY_BUCKET=registry-data
 OPEN_BALENA_JWT_SECRET=
 
 REACT_APP_OPEN_BALENA_POSTGREST_URL=https://postgrest.my-open-balena-server.com
