+++ server/util/s3/bucketNames.js
@@ -1,0 +1,5 @@
+require('dotenv').config();
+
+module.exports = {
+  registry: process.env.OPEN_BALENA_S3_REGISTRY_BUCKET || 'registry-data'
+}
