+++ server/util/s3/index.js
@@ -1,0 +1,4 @@
+module.exports = {
+  s3Client: require('./client'),
+  bucketNames: require('./bucketNames')
+}
